Advisor/requirements.txt
google-genai
watchdog
PyQt6
python-dotenv